feat: exclude NaN from summation, min, max
For pragmatic reasons, we decided to have sources send NaN values for non-responsive devices. Those values are not stored in the DB to not skew the aggregation. To do the same in combined metrics but also avoid the combined metrics being unavailable during the unavailability of one input metric, we want to drop them from certain operations. Specifically this is addition/summation and min/max.